refactor(cli): move retrieval formatters to adoc-cli presentation layer
