+++ crates/adoc-cli/src/presentation/mod.rs
@@ -1,0 +1,7 @@
+pub(crate) mod json;
+pub(crate) mod plain;
+pub(crate) mod port;
+
+pub(crate) use json::JsonPresenter;
+pub(crate) use plain::PlainPresenter;
+pub(crate) use port::ExplainPresenter;
