+++ crates/adoc-core/src/application/mod.rs
@@ -3,4 +3,3 @@ pub(crate) mod compile;
 pub(crate) mod resolve_knowledge_objects;
 pub(crate) mod resolve_object_references;
 pub(crate) mod retrieval;
-pub(crate) mod retrieval_format;
