+++ crates/adoc-core/tests/retrieval.rs
@@ -3,9 +3,9 @@ use std::path::PathBuf;
 
 use adoc_core::{
     AgentJsonDocument, AgentJsonObject, AgentJsonRelations, AgentJsonSourceSpan, DiagnosticCode,
-    JsonRetrievalFormatter, RetrievalEnvelope, RetrievalFormatter, RetrievalInput, RetrievalMatch,
-    RetrievalRecord, RetrievalSession, RetrievalSource, SearchFilters, SearchMode, SearchQuery,
-    SearchResult, TextRetrievalFormatter, explain_object, load_retrieval_session, search,
+    RetrievalEnvelope, RetrievalInput, RetrievalMatch, RetrievalRecord, RetrievalSession,
+    RetrievalSource, SearchFilters, SearchMode, SearchQuery, SearchResult, explain_object,
+    load_retrieval_session, search,
 };
 
 fn fixture_path(relative: &str) -> PathBuf {
@@ -1042,222 +1042,4 @@ fn load_retrieval_session_rejects_duplicate_object_ids_inside_artifact() {
         result.diagnostics[0].object_id.as_deref(),
         Some("billing.duplicate")
     );
-}
-
-#[test]
-fn text_retrieval_formatter_renders_statement_body_and_sorted_fields() {
-    let envelope = RetrievalEnvelope::new(
-        vec![RetrievalRecord {
-            id: "billing.policy".to_string(),
-            kind: "decision".to_string(),
-            status: Some("accepted".to_string()),
-            owner: Some("architecture".to_string()),
-            verified_at: None,
-            body: "Refund policy is ledger-backed.\nManual credits are exceptions.".to_string(),
-            source: RetrievalSource {
-                path: "docs/decisions.adoc".to_string(),
-                line: 7,
-                column: 1,
-            },
-            evidence: std::collections::BTreeMap::new(),
-            fields: std::collections::BTreeMap::from([
-                ("scope".to_string(), "refunds".to_string()),
-                ("decided_by".to_string(), "architecture".to_string()),
-            ]),
-            relations: AgentJsonRelations::default(),
-            search_match: None,
-        }],
-        Vec::new(),
-    );
-
-    let text = TextRetrievalFormatter
-        .render(&envelope)
-        .expect("text retrieval render succeeds");
-
-    assert_eq!(
-        text,
-        concat!(
-            "Object: billing.policy\n",
-            "Kind: decision\n",
-            "Status: accepted\n",
-            "Owner: architecture\n",
-            "\n",
-            "Statement:\n",
-            "Refund policy is ledger-backed.\n",
-            "Manual credits are exceptions.\n",
-            "\n",
-            "Fields:\n",
-            "- decided_by: architecture\n",
-            "- scope: refunds\n",
-            "\n",
-            "Source: docs/decisions.adoc:7:1\n",
-        )
-    );
-}
-
-#[test]
-fn text_retrieval_formatter_renders_each_relation_target_on_its_own_line() {
-    let envelope = RetrievalEnvelope::new(
-        vec![RetrievalRecord {
-            id: "billing.policy".to_string(),
-            kind: "decision".to_string(),
-            status: Some("accepted".to_string()),
-            owner: None,
-            verified_at: None,
-            body: "Refund policy is ledger-backed.".to_string(),
-            source: RetrievalSource {
-                path: "docs/decisions.adoc".to_string(),
-                line: 7,
-                column: 1,
-            },
-            evidence: std::collections::BTreeMap::new(),
-            fields: std::collections::BTreeMap::new(),
-            relations: AgentJsonRelations {
-                depends_on: vec![
-                    "billing.credits.ledger-source".to_string(),
-                    "billing.refunds.audit-required".to_string(),
-                ],
-                supersedes: vec![
-                    "billing.refunds.manual-credit".to_string(),
-                    "billing.refunds.email-approval".to_string(),
-                ],
-                related_to: vec![
-                    "billing.credits.decrement-after-success".to_string(),
-                    "billing.credits.reconciliation".to_string(),
-                ],
-            },
-            search_match: None,
-        }],
-        Vec::new(),
-    );
-
-    let text = TextRetrievalFormatter
-        .render(&envelope)
-        .expect("text retrieval render succeeds");
-    let relations = text
-        .split_once("Relations:\n")
-        .expect("relations block is rendered")
-        .1;
-
-    assert_eq!(
-        relations,
-        concat!(
-            "- depends_on: billing.credits.ledger-source\n",
-            "- depends_on: billing.refunds.audit-required\n",
-            "- supersedes: billing.refunds.manual-credit\n",
-            "- supersedes: billing.refunds.email-approval\n",
-            "- related_to: billing.credits.decrement-after-success\n",
-            "- related_to: billing.credits.reconciliation\n",
-        )
-    );
-}
-
-#[test]
-fn text_retrieval_formatter_renders_glossary_kind_metadata() {
-    let envelope = RetrievalEnvelope::new(
-        vec![RetrievalRecord {
-            id: "billing.credits".to_string(),
-            kind: "glossary".to_string(),
-            status: None,
-            owner: None,
-            verified_at: None,
-            body: "Credits are account balance units.".to_string(),
-            source: RetrievalSource {
-                path: "docs/glossary.adoc".to_string(),
-                line: 4,
-                column: 1,
-            },
-            evidence: std::collections::BTreeMap::new(),
-            fields: std::collections::BTreeMap::from([(
-                "canonical".to_string(),
-                "billing credit".to_string(),
-            )]),
-            relations: AgentJsonRelations::default(),
-            search_match: None,
-        }],
-        Vec::new(),
-    );
-
-    let text = TextRetrievalFormatter
-        .render(&envelope)
-        .expect("text retrieval render succeeds");
-
-    assert!(text.contains("Kind: glossary\n"));
-    assert!(text.contains("Fields:\n- canonical: billing credit\n"));
-}
-
-#[test]
-fn text_retrieval_formatter_renders_unknown_evidence_keys_after_known_order() {
-    let envelope = RetrievalEnvelope::new(
-        vec![RetrievalRecord {
-            id: "billing.credits".to_string(),
-            kind: "claim".to_string(),
-            status: Some("verified".to_string()),
-            owner: None,
-            verified_at: None,
-            body: "Credits decrement after successful payment.".to_string(),
-            source: RetrievalSource {
-                path: "docs/billing.adoc".to_string(),
-                line: 9,
-                column: 1,
-            },
-            evidence: std::collections::BTreeMap::from([
-                ("artifact".to_string(), "ledger.csv".to_string()),
-                ("reviewed_by".to_string(), "risk".to_string()),
-                ("source".to_string(), "ledger".to_string()),
-                ("z_probe".to_string(), "trace".to_string()),
-            ]),
-            fields: std::collections::BTreeMap::new(),
-            relations: AgentJsonRelations::default(),
-            search_match: None,
-        }],
-        Vec::new(),
-    );
-
-    let text = TextRetrievalFormatter
-        .render(&envelope)
-        .expect("text retrieval render succeeds");
-
-    assert!(text.contains(concat!(
-        "Evidence:\n",
-        "- source: ledger\n",
-        "- reviewed_by: risk\n",
-        "- artifact: ledger.csv\n",
-        "- z_probe: trace\n",
-    )));
-}
-
-#[test]
-fn json_retrieval_formatter_preserves_envelope_shape() {
-    let envelope = RetrievalEnvelope::new(
-        vec![RetrievalRecord {
-            id: "billing.credits".to_string(),
-            kind: "claim".to_string(),
-            status: Some("verified".to_string()),
-            owner: None,
-            verified_at: None,
-            body: "Credits decrement after successful payment.".to_string(),
-            source: RetrievalSource {
-                path: "docs/billing.adoc".to_string(),
-                line: 9,
-                column: 1,
-            },
-            evidence: std::collections::BTreeMap::new(),
-            fields: std::collections::BTreeMap::new(),
-            relations: AgentJsonRelations::default(),
-            search_match: None,
-        }],
-        Vec::new(),
-    );
-
-    let rendered = JsonRetrievalFormatter
-        .render(&envelope)
-        .expect("JSON retrieval render succeeds");
-    let expected = serde_json::to_string_pretty(&envelope).expect("envelope serializes");
-
-    assert_eq!(rendered, expected);
-    let value: serde_json::Value = serde_json::from_str(&rendered).expect("rendered JSON parses");
-    assert_eq!(value["schema_version"], "adoc.retrieval.v0");
-    assert!(value["records"][0].get("match").is_none());
-    assert!(value["records"][0].get("retrieval").is_none());
 }
